+++ src/main/kotlin/blackjack/entity/Card.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.entity
 
 data class Card(val suit: Suit, val rank: Rank)
