+++ src/main/kotlin/blackjack/entity/Hand.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.entity
 
 class Hand(val cards: MutableList<Card> = mutableListOf()) {
     fun addCard(card: Card) {
