+++ src/test/kotlin/blackjack/entity/DeckTest.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.entity
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
